Optimizing SQL Queries:
.select_related() and .prefetch_related().
Remember the direction each of these goes. select_related is used on the model when you have the ForeignKey field. prefetch_related is used on the model that's related to by the ForeignKey field.
prefetch_related won't always reduce the number of queries. It helps to prevent extra queries being run in your templates, though, by fetching and attaching the data before the template is ever rendered.
select_related, when used correctly, can drastically reduce the number of queries you run.